feat: add search-item interface
