+++ YouTube-client-app/src/app/search/search-item/search-item.component.spec.ts
@@ -8,9 +8,8 @@ describe('SearchItemComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [SearchItemComponent]
-    })
-    .compileComponents();
+      imports: [SearchItemComponent],
+    }).compileComponents();
 
     fixture = TestBed.createComponent(SearchItemComponent);
     component = fixture.componentInstance;
